requirements.txt
streamlit==0.84.0
torch==1.9.0
torchvision==0.10.0
Pillow==8.0.1
requests==2.21.0